Fix a potential data loss bug
When reorganising existing data there is a bug related to the order in
which moves are performed that can result in files overwriting each
other thus losing data as a side effect of the sequence of file
movements.  This commit adds in conflict detection and an intermediate
rename of affected source files to avoid this problem.